Support more than one set of genl commands.
In preparation for supporting the upstream kernel set of MPTCP path
management generic netlink commands, a new mptcpd_pm_cmd_ops struct
was created that contains the operations corresponding to the commands
supported by both the upstream kernel and the multipath-tcp.org
kernel.  One set of commands, i.e. a struct mptcpd_pm_cmd_ops
instance, will be available for each kernel.

Besides introducing the mptcpd_pm_cmd_ops struct, this set of changes
moves the multipath-tcp.org kernel generic netlink command
implementation out of libmptcpd and into the mptcpd program.
Libmptcpd will indirectly call those commands through the cmd_ops
field in the mptcpd_pm struct without having to know about the
underlying implementation details.

Some command arguments were re-ordered to give preference to the
upstream kernel required set of arguments in cases where the commands
overlapped between the two kernels, i.e. add_addr and remove_addr.
